# FSSR: Compute node center incrementally in Octree::influence_query()