Speed-up over Octree::node_center_and_size() on citywall-20140923 (L2) dataset:
1x Intel Core i7-3930K (6 cores, 12 threads)     : ~1.43x (~7086.1s -> ~4962.4s)
2x Intel Xeon E5-2687W v3 (2x8 cores, 32 threads): ~1.41x (~3099.0s -> ~2197.6s)

Speed-up over Octree::node_center_and_size() on achteck_turm-20130528 (L2) dataset:
1x Intel Core i7-3930K (6 cores, 12 threads)     : ~2.06x (~1452.2s -> ~705.3s)
2x Intel Xeon E5-2687W v3 (2x8 cores, 32 threads): ~1.89x (~ 644.8s -> ~341.7s)